requirements.txt
requests
google-generativeai